requirements.txt
# Needed for OpenUpgrade
odoorpc==0.7.0
git+https://github.com/OCA/openupgradelib.git@master
#para facturacion electronica
suds-community